requirements.txt
asgiref==3.2.10
Django==3.0.7
django-crispy-forms==1.9.1
django-filter==2.3.0
django-json-field==0.5.7
django-templated-mail==1.1.1
djangorestframework==3.11.0
djoser==2.0.3
drf-nested-routers==0.91
gunicorn==20.0.4
importlib-metadata==1.7.0
Markdown==3.2.2
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
uuid==1.30
zipp==3.1.0
django-heroku
